backend/pyproject.toml
[tool.poetry]
name = "backend-finmanagment"
version = "0.1.0"
description = "Backend of the financial managment api"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
python-dotenv = "^1.1.0"
pytest = "^8.3.5"
ruff = "^0.11.11"
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
